requirements.txt
Flask-r
Flask-SQLAlchemy
Flask-Migrate
Flask-Login

